feat: Enhanced dashboard with improved data visualization and period filtering
- Add comprehensive period filtering with day/week/month/quarter/year support
- Implement cumulative cash flow tracking with detailed operation display
- Create new chart components: IncomeExpenseChart, WeeklyFlowChart, AccountBalancesChart
- Add RecentOperationsTable for better operation visibility
- Enhance data aggregation utilities with smart interval creation
- Add comprehensive test coverage for new functionality
- Improve UI/UX with custom tooltips and better period navigation
- Add shared utilities for date handling and period management
- Optimize chart data formatting and operation aggregation
