+++ apps/api/src/modules/reports/dds/dds.service.ts
@@ -1,5 +1,5 @@
 import prisma from '../../../config/db';
-import { getMonthKey, getMonthsBetween } from '../utils/date';
+import { getMonthKey, getMonthsBetween } from '@fin-u-ch/shared';
 import { cacheReport, getCachedReport, generateCacheKey } from '../utils/cache';
 
 export interface DDSParams {
