+++ apps/api/src/modules/reports/planfact/planfact.service.ts
@@ -1,5 +1,5 @@
 import prisma from '../../../config/db';
-import { getMonthKey } from '../utils/date';
+import { getMonthKey } from '@fin-u-ch/shared';
 import { cacheReport, getCachedReport, generateCacheKey } from '../utils/cache';
 import plansService from '../../plans/plans.service';
