+++ packages/shared/package.json
@@ -17,6 +17,7 @@
     "ts-jest": "^29.1.0"
   },
   "dependencies": {
-    "zod": "^4.1.12"
+    "zod": "^4.1.12",
+    "date-fns": "^3.6.0"
   }
 }
