+++ packages/shared/src/index.ts
@@ -8,6 +8,7 @@ export * from './types/catalogs';
 export * from './types/operations';
 export * from './types/reports';
 export * from './types/notifications';
+export * from './types/period';
 
 // Schemas
 export * from './schemas/notifications';
@@ -18,3 +19,5 @@ export * from './constants/config';
 
 // Utilities
 export * from './lib/currencies';
+export * from './lib/period';
+export * from './lib/date';
