+++ packages/shared/src/types/period.ts
@@ -1,0 +1,12 @@
+export type PeriodFormat = 'day' | 'week' | 'month' | 'quarter' | 'year';
+
+export interface PeriodRange {
+  from: string;
+  to: string;
+}
+
+export interface Interval {
+  start: Date;
+  end: Date;
+  label: string;
+}
